pbdNCDF4-package {pbdNCDF4}	R Documentation

Programming with Big Data – Interface to Parallel Unidata NetCDF4 Format Data Files

Description

This package adds collective parallel read and write capability to the R package ncdf4 version 1.8. Typical use is as a parallel NetCDF4 file reader in SPMD style programming. Each R process reads and writes its own data in a synchronized collective mode, resulting in faster parallel performance. Performance improvement is conditional on a parallel file system.

Details

Package: 	 pbdNCDF4

Type: 	 Package

License: 	 GPL

LazyLoad: 	 yes

The parallel HDF5 and NetCDF4 (version 4.1 or later) libraries are prerequisite, then this package will link with them and improve the I/O performance via three mainly functions nc_creat_par, nc_open_par, and nc_var_par_access added to the original ncdf4 (1.8) package to enable collective reading and writing.

If only serial NetCDF4 and HDF5 libraries are available, then this package will function exactly the same as ncdf4. All functionalities of ncdf4 (version 1.8) are followed.

Author(s)

George Ostrouchov, Pragneshkumar Patel pragnesh@utk.edu, Wei-Chen Chen, Drew Schmidt, and David Pierce.

References

The HDF Group: http://www.hdfgroup.org/HDF5/

Unidata: http://www.unidata.ucar.edu/software/netcdf/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

nc_create_par, nc_open_par, nc_var_par_access, and ncdump.

Examples

## Not run: 
### Under command mode, run the demo with 2 processors by
### (Use Rscript.exe for windows system)
mpiexec -np 2 Rscript -e "demo(ncwrite,'pbdNCDF4',ask=F,echo=F)"
mpiexec -np 2 Rscript -e "demo(ncread,'pbdNCDF4',ask=F,echo=F)"
mpiexec -np 2 Rscript -e "demo(ncdump,'pbdNCDF4',ask=F,echo=F)"

## End(Not run)

[Package pbdNCDF4 version 0.1-4 Index]